feat: revise tests, add an endpoint, use loader in the router
The test is now using anyio plugin for test with direct database access.
The client is now a pytest's fixture. Test database is now an
in-memory one.

On client side, API functions are now centralized. The data is passed to
the element of the route via router's data loader instead of manually
fetch the data in the component.